backend/requirements.txt
fastapi==0.109.1
uvicorn[standard]==0.24.0
websockets==12.0
pyneurosdk2>=1.0.15
python-multipart==0.0.18
pydantic==2.5.0
pyem-st-artifacts
